+++ app/src/main/java/com/ikvych/cocktail/presentation/fragment/CustomViewFragment.kt
@@ -2,14 +2,11 @@ package com.ikvych.cocktail.presentation.fragment
 
 import android.os.Bundle
 import android.text.TextUtils
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
-import androidx.databinding.ViewDataBinding
 import com.ikvych.cocktail.R
 import com.ikvych.cocktail.databinding.FragmentCustomViewBinding
 import com.ikvych.cocktail.presentation.fragment.base.BaseFragment
+import com.ikvych.cocktail.util.widget.custom.LinearLayoutWithChildLimiter
 import com.ikvych.cocktail.viewmodel.base.BaseViewModel
 import kotlinx.android.synthetic.main.fragment_custom_view.*
 import kotlin.reflect.KClass
@@ -25,6 +22,8 @@ class CustomViewFragment : BaseFragment<BaseViewModel, FragmentCustomViewBinding
         artv_test.maxLines = 2
         artv_test.ellipsize = TextUtils.TruncateAt.END
         artv_test.text = text
+        val lp = tv_test.layoutParams as LinearLayoutWithChildLimiter.LinearLayoutWithChildLimiterLayoutParams
+
     }
 
     companion object {
